% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{ffreg}
\alias{ffreg}
\title{Estimation function for the classic function-on-function linear regression model}
\usage{
ffreg(Y, X, grd, rho = NULL, rho_rng = c(0, 100))
}
\arguments{
\item{Y}{Outcome}

\item{X}{Predictor}

\item{grd}{Grid}

\item{rho}{Smoothing parameter. If left unspecified (rho = NULL), then rho is
determined by Generalized Cross Validation (GCV).}

\item{rho_rng}{The range c(min(rho_rng), max(rho_rng)) is used for finding
the GCV-optimal smoothing parameter rho, if rho = NULL.}
}
\description{
Estimation function for the classic function-on-function linear regression model
}
