% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{ppfunreg}
\alias{ppfunreg}
\title{Estimation function}
\usage{
ppfunreg(Y, X, grd, rho = NULL, rho_rng = NULL)
}
\arguments{
\item{Y}{Outcome}

\item{X}{Predictor}

\item{grd}{Grid}

\item{rho}{Smoothingparameter}
}
\description{
Estimation function
}
