% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{ppfunreg2}
\alias{ppfunreg2}
\title{Estimation function}
\usage{
ppfunreg2(Y, X, grd, rho = NULL, rho_rng = c(0, 100))
}
\arguments{
\item{Y}{Outcome. A matrix with ncol(Y) = 'sample size' and nrow(Y) = 'number of grid 'grd' points'}

\item{X}{Predictor. A matrix with ncol(X) = 'sample size' and nrow(X) = 'number of grid 'grd' points'}

\item{grd}{Grid between 0 and 1.}

\item{rho}{Smoothing parameter. If left unspecified (rho = NULL), then rho is
determined by Generalized Cross Validation (GCV).}

\item{rho_rng}{The range c(min(rho_rng), max(rho_rng)) is used for finding
the GCV-optimal smoothing parameter rho, if rho = NULL.}
}
\description{
Estimation function
}
